+++ src/main/java/roomescape/controller/page/UserPageController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.page;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
